Split off management page controller
